Fixed performance issues with monocular and monocular-inertial
I was having uncharacteristically extremely bad performance from both
monocular and monocular-inertial modes when running orb slam. I fixed
this by changing the resolution of the realsense camera to 640x480x30 in
the launch file and in the orbslam3 config file.

Orbslam3 seems to not like the default resolution of the camera,
1280x720x30.