Fixed issue with how I was packaging imu data
As a remnant from when I was trying to post process the images and imu
data to improve performance, I was packaging my imu data into two
buffers. One "temporary" buffer would be made, and then packaged into
another primary buffer that was supposed to be full of these temporary
buffers. I'm pretty sure that since I changed to real time, the
temporary buffers I was making combined with how I was processing the
images was creating a strange scenario with too little or strange imu
data that was causing orbslam3 to freak out. After this change, there
are still errors and problems initializing the imu data it seems, but
the program doesn't crash anymore.