Attempting to grab pose graph and use it for laser_scan
I want to get the current pose of the camera from orb_slam3 corrected by
the imu and use it as the base_footprint frame and transform that frame
around. I'll use the base_footprint as the frame_id of the laser scan so
that it will work correctly. Almost done