laser scan, current key frame point cloud, and pose graph are aligned
Had to do a decent amount of fiddling, and there's a hard coded angle
adjustment in one of the functions. At some point, I should revisit this
and get it right. I couldn't figure out why the laser_scan was off 90
degrees from the point cloud, it's some math remnant of how I'm
converting the 3d point cloud to a 2d laser scan