set up basic libpointmatcher node
I also looked at the point cloud generated from the velodyne sensor, and
I don't think that libpointmatcher will be able to handle the
differences between the two point clouds. orb-slam3 is so much sparser
and the points are very randomly places around. Plus, the camera from
the wheelchair is very low down. I don't think it's point clouds are
scaled either, because they are pretty large.

I have two ideas right now. One is to perform monocular orbslam using
the top third of the grayscale camera feed from the wheelchair. This
would give me an unscaled 3d point cloud that I could try to  match with
the one I made. My other option is to make an occupancy grid using my
vio slam data, and then to run slam_toolbox on the wheelchair and try to
perform 2d point cloud matching on the occupancy grids, or find some
other way to localize with them. I'm thinking that the easiest thing to
try right now is to use icp to match the occupancy grids, and that's
what I'm going to try first.