2D point matching between occupancy grids is working
I didn't end up having to use the filter that I created, which is good
because it wasn't working anyways. Now that this is working, I need to
go to the lab and drive the wheel chair around in the same place that I
have camera data. Then I can try and match those maps and we'll really
be done. After that's done I need to clean up the code in a major way,
things have gotten messy over the past few days.