Massive overhaul of the architecture of the project
I made the mapping launch file more exclusively for mapping, and I made
it so it can be run alongside the localize launch file and basically
plug into it. The transform that is calculated by libpointmatcher is
used to transform the a frame for the live_map to be the child of,
instead of transforming the points in the point clloud themselves and
leaving the tf static to the map.